monitor: fix notifications when connecting to the network or pluggin a disk
We can't use the reactive_check variable directly, because that is
set and reset synchronously around the prepare_run() calls, but
the code that checks it runs after a yield (so after an async call
and at least one mainloop iteration). Instead, use a local temporary
that vala will know to store and retrieve.