Hide cloud backends by default.
GNOME Online Accounts is now the preferred cloud provider. This lets us
standardize on gvfs and they are generally more user-friendly than signing
up for a S3 account or the like.

Cloud providers are still available for use and will show up in the UI if
the user was using one before or if they manually set it in gsettings.

We also show all cloud providers in the drop down when restoring (since who
knows where the user had their data at that point).