Fix settings being marked read-only when started via --restore
We used to keep a global read-only setting if we were doing a restore.
But now with dbus-activation and one central process, that didn't make
any sense. And we could stick the whole process in read-only settings
mode if we were started with with --restore.

Instead of a global settings table, we now have ConfigLocation generate
a Backend object that gets passed down to the Operation objects.
Likewise we are careful in various places to use the same Settings
objects.