monitor: limit memory usage of deja-dup-monitor
We've had reports of deja-dup-monitor taking all the system memory.
But I haven't been able to reproduce. In an attempt to prevent the
monitor from dragging down the system, set a limit on memory use
by the autostart monitor process using ulimit.

This needs continued watching to see if we can fix this for real.
And to make sure we haven't made anything worse with this
workaround.

(Also in this commit, we rename the file and move it in-source.)

https://launchpad.net/bugs/1302416