BackendRemote: Make "SMB host is not available" message nicer
GVFS gives us back "Invalid argument" which isn't very user friendly.
So we try to detect that situation and promote it to a nicer message.

Also fix testing whether the host is reachable when already mounted,
instead just checking if we are connected to the network at all.
SSH configs and such may fool us into thinking a host nickname isn't
reachable.  If it's mounted and we are connected, that should be
enough.